ISSUE #3878 - move ellipses functions into ellipses components
