+++ frontend/src/v5/ui/routes/dashboard/projects/containers/containersList/containerListItem/containerEllipsisMenu/containerEllipsisMenu.component.tsx
@@ -21,26 +21,58 @@ import { ContainersActionsDispatchers, DialogsActionsDispatchers } from '@/v5/se
 import { EllipsisMenu } from '@controls/ellipsisMenu/ellipsisMenu.component';
 import { EllipsisMenuItem } from '@controls/ellipsisMenu/ellipsisMenuItem/ellipsisMenutItem.component';
 import { canUploadToBackend } from '@/v5/store/containers/containers.helpers';
-import { viewerRoute } from '@/v5/services/routing/routing';
+import { prefixBaseDomain, viewerRoute } from '@/v5/services/routing/routing';
 import { DashboardParams } from '@/v5/ui/routes/routes.constants';
 import { uploadToContainer } from '../../../uploadFileForm/uploadFileForm.helpers';
+import { ContainerSettingsModal } from '../../../containerSettingsModal/containerSettingsModal.component';
 
 type ContainerEllipsisMenuProps = {
 	selected: boolean,
 	container: IContainer,
 	onSelectOrToggleItem: (id: string) => void,
-	openShareModal: () => void,
-	openContainerSettings: () => void,
 };
 
 export const ContainerEllipsisMenu = ({
 	selected,
 	container,
 	onSelectOrToggleItem,
-	openShareModal,
-	openContainerSettings,
 }: ContainerEllipsisMenuProps) => {
 	const { teamspace, project } = useParams<DashboardParams>();
+
+	const onClickShare = () => {
+		const link = prefixBaseDomain(viewerRoute(teamspace, project, container));
+		const subject = formatMessage({ id: 'shareModal.container.subject', defaultMessage: 'container' });
+		const title = formatMessage({ id: 'shareModal.container.title', defaultMessage: 'Share Container' });
+
+		DialogsActionsDispatchers.open('share', {
+			name: container.name,
+			subject,
+			title,
+			link,
+		});
+	};
+
+	// eslint-disable-next-line max-len
+	const onClickSettings = () => DialogsActionsDispatchers.open(ContainerSettingsModal, { containerId: container._id });
+
+	const onClickDelete = () => DialogsActionsDispatchers.open('delete', {
+		name: container.name,
+		onClickConfirm: () => new Promise<void>(
+			(accept, reject) => {
+				ContainersActionsDispatchers.deleteContainer(
+					teamspace,
+					project,
+					container._id,
+					accept,
+					reject,
+				);
+			},
+		),
+		message: formatMessage({
+			id: 'deleteModal.container.message',
+			defaultMessage: 'By deleting this Container your data will be lost permanently and will not be recoverable.',
+		}),
+	});
 
 	return (
 		<EllipsisMenu selected={selected}>
@@ -93,38 +125,21 @@ export const ContainerEllipsisMenu = ({
 					id: 'containers.ellipsisMenu.shareContainer',
 					defaultMessage: 'Share Container',
 				})}
-				onClick={openShareModal}
+				onClick={onClickShare}
 			/>
 			<EllipsisMenuItem
 				title={formatMessage({
 					id: 'containers.ellipsisMenu.settings',
 					defaultMessage: 'Settings',
 				})}
-				onClick={openContainerSettings}
+				onClick={onClickSettings}
 			/>
 			<EllipsisMenuItem
 				title={formatMessage({
 					id: 'containers.ellipsisMenu.delete',
 					defaultMessage: 'Delete',
 				})}
-				onClick={() => DialogsActionsDispatchers.open('delete', {
-					name: container.name,
-					onClickConfirm: () => new Promise<void>(
-						(accept, reject) => {
-							ContainersActionsDispatchers.deleteContainer(
-								teamspace,
-								project,
-								container._id,
-								accept,
-								reject,
-							);
-						},
-					),
-					message: formatMessage({
-						id: 'deleteModal.container.message',
-						defaultMessage: 'By deleting this Container your data will be lost permanently and will not be recoverable.',
-					}),
-				})}
+				onClick={onClickDelete}
 			/>
 		</EllipsisMenu>
 	);
